Project Creation
I've managed to get QT to work correctly despite hours of issues. At this stage, I have a very BASIC project set up, with a GUI that somewhat resembles the intended final product.

At this rate, this project won't take long, other than the need for past encryption to be recognizable. That's the only issue I can think of with the encryption key limited for 5 characters. Maybe add a workaround..?